chore: clear the clippy backlog, so a real warning is visible again
Ten lints across six files, none of them from the capture seam and none
of them behaviour-changing: a missing `Default`, a `Default::default()`
followed by field assignment, two no-op `u32 as u32` casts, three
`if`s that edition 2024 lets fold into let-chains, a `loop`/let-else
that is a `while let`, a `&PathBuf` parameter that only ever needed
`&Path`, and a heap `vec!` for a 32-byte const-sized mask.

They accumulated because the gate that would have caught them was being
run as `-p trix-cli`. `cargo clippy --workspace --all-targets -- -D
warnings` is clean now, which is the point: nine standing warnings are a
place for the tenth to hide.

+++ crates/trix-core/src/encode/mf.rs
@@ -38,7 +38,7 @@ use crate::capture::audio::{CHANNELS, ENCODER_BLOCK_ALIGN, SAMPLE_RATE};
 use crate::config::RateControl;
 use crate::encode::convert::VideoConverter;
 
-const MF_VERSION: u32 = ((MF_SDK_VERSION as u32) << 16) | MF_API_VERSION as u32;
+const MF_VERSION: u32 = (MF_SDK_VERSION << 16) | MF_API_VERSION;
 
 /// Media Foundation startup is process-wide; do it exactly once. `MFShutdown`
 /// is deliberately skipped — the OS reclaims everything at process exit.
